getting id of inserted row bug fixed using couroutine

+++ app/src/main/java/com/gaurav/budgetplanner/features/reminder/presentation/adapter/ReminderAdapter.kt
@@ -52,7 +52,7 @@ class ReminderAdapter(): RecyclerView.Adapter<ReminderAdapter.MyViewHolder>() {
 
 
         holder.binding.btnRemind.setOnCheckedChangeListener { _, isChecked ->
-            onSwitchChange?.invoke(eachItem.id,isChecked,eachItem)
+            onSwitchChange?.invoke(eachItem.id.toInt(),isChecked,eachItem)
         }
     }
